+++ src/main/java/insane96mcp/iguanatweaksreborn/data/generator/ISOItemTagsProvider.java
@@ -38,7 +38,6 @@ public class ISOItemTagsProvider extends ItemTagsProvider {
     public static final TagKey<Item> STONE_HAND_EQUIPMENT = ISOItemTagsProvider.create("equipment/hand/stone");
     public static final TagKey<Item> FLINT_HAND_EQUIPMENT = ISOItemTagsProvider.create("equipment/hand/flint");
     public static final TagKey<Item> GOLDEN_HAND_EQUIPMENT = ISOItemTagsProvider.create("equipment/hand/golden");
-    public static final TagKey<Item> GOLDEN_TOOL_EQUIPMENT = ISOItemTagsProvider.create("equipment/hand/golden/tools");
     public static final TagKey<Item> LEATHER_ARMOR_EQUIPMENT = ISOItemTagsProvider.create("equipment/armor/leather");
 
     public ISOItemTagsProvider(PackOutput packOutput, CompletableFuture<HolderLookup.Provider> completableFuture, CompletableFuture<TagLookup<Block>> tagLookupCompletableFuture, String modId, @Nullable ExistingFileHelper existingFileHelper) {