+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/plugin/ModLoadedPlugin.java
@@ -24,7 +24,8 @@ public class ModLoadedPlugin implements IMixinConfigPlugin {
             .put("insane96mcp.iguanatweaksreborn.mixin.integration.autumnity.CookedTurkeyBlockMixin", () -> LoadingModList.get().getModFileById("autumnity") != null)
             .put("insane96mcp.iguanatweaksreborn.mixin.integration.autumnity.TurkeyBlockMixin", () -> LoadingModList.get().getModFileById("autumnity") != null)
             .put("insane96mcp.iguanatweaksreborn.mixin.integration.autumnity.PancakeBlockMixin", () -> LoadingModList.get().getModFileById("autumnity") != null)
-            .put("insane96mcp.iguanatweaksreborn.mixin.integration.farmersdelight.PieBlockMixin", () -> LoadingModList.get().getModFileById("farmersdelight") != null)
+            //.put("insane96mcp.iguanatweaksreborn.mixin.integration.farmersdelight.PieBlockMixin", () -> LoadingModList.get().getModFileById("farmersdelight") != null)
+            .put("insane96mcp.iguanatweaksreborn.mixin.integration.cavernsandchasms.CCEntityTypesMixin", () -> LoadingModList.get().getModFileById("caverns_and_chasms") != null)
             .build();
     @Override
     public void onLoad(String mixinPackage) {